pyproject.toml
[tool.poetry]
name = "data-bureau"
version = "0.1.0"
description = ""
authors = ["NerdPraise <praiseajayi2@gmail.com>"]
readme = "README.md"
packages = [{include = "data_bureau"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-allauth = "^0.52.0"
black = "^23.1.0"
isort = "^5.12.0"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"
django-phonenumber-field = "^7.0.2"
coreapi = "^2.3.3"
pyyaml = "^6.0"
drf-spectacular = "^0.25.1"
python-dotenv = "^0.21.1"
rsa = "^4.9"


[tool.poetry.group.dev.dependencies]
phonenumbers = "^8.13.6"
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
